# public/openapi.yaml
openapi: 3.0.0
info:
  title: Leemage API
  version: 1.0.0
  description: Leemage 파일 관리 플랫폼 API. 프로젝트 단위로 파일을 관리하고 이미지 변환 기능을 제공합니다.
servers:
  - url: /api/v1
    description: API v1 (API Key 인증)
tags:
  - name: Projects
    description: 프로젝트 관리 API
  - name: Files
    description: 파일 업로드 및 관리 API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "API 키를 사용한 인증. 'Authorization: Bearer <API_KEY>' 형식으로 전달"
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: 에러 메시지
          example: 요청 처리 중 오류가 발생했습니다.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: 필드별 유효성 검사 에러
          example:
            name:
              - 필수 항목입니다.
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: 성공 메시지
          example: 작업이 완료되었습니다.
      required:
        - message
    VariantOption:
      type: object
      properties:
        sizeLabel:
          type: string
          enum:
            - original
            - 300x300
            - 800x800
            - 1920x1080
          description: 이미지 크기 레이블
          example: original
        format:
          type: string
          enum:
            - png
            - jpeg
            - avif
            - webp
          description: 이미지 포맷
          example: webp
      required:
        - sizeLabel
        - format
    ImageVariantData:
      type: object
      properties:
        url:
          type: string
          description: 이미지 URL
          example: https://objectstorage.ap-seoul-1.oraclecloud.com/...
        width:
          type: number
          description: 이미지 너비 (px)
          example: 1920
        height:
          type: number
          description: 이미지 높이 (px)
          example: 1080
        size:
          type: number
          description: 파일 크기 (bytes)
          example: 102400
        format:
          type: string
          description: 이미지 포맷
          example: webp
        label:
          type: string
          description: 크기 레이블
          example: original
      required:
        - url
        - width
        - height
        - size
        - format
        - label
    PresignRequest:
      type: object
      properties:
        fileName:
          type: string
          minLength: 1
          description: 업로드할 파일의 이름
          example: image.jpg
        contentType:
          type: string
          minLength: 1
          description: 파일의 MIME 타입
          example: image/jpeg
        fileSize:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: 파일 크기 (bytes)
          example: 102400
      required:
        - fileName
        - contentType
        - fileSize
    PresignResponse:
      type: object
      properties:
        presignedUrl:
          type: string
          description: OCI에 직접 업로드할 수 있는 Presigned URL
          example: https://objectstorage.ap-seoul-1.oraclecloud.com/p/...
        objectName:
          type: string
          description: OCI 객체 경로
          example: clq1234abcd/file5678efgh.jpg
        objectUrl:
          type: string
          description: 업로드 완료 후 객체 URL
          example: https://objectstorage.ap-seoul-1.oraclecloud.com/n/.../o/...
        fileId:
          type: string
          description: 생성된 파일 ID
          example: file5678efgh
        expiresAt:
          type: string
          description: Presigned URL 만료 시간 (ISO 8601)
          example: 2023-01-01T00:15:00.000Z
      required:
        - presignedUrl
        - objectName
        - objectUrl
        - fileId
        - expiresAt
    ConfirmRequest:
      type: object
      properties:
        fileId:
          type: string
          minLength: 1
          description: presign 응답에서 받은 파일 ID
          example: file5678efgh
        objectName:
          type: string
          minLength: 1
          description: presign 응답에서 받은 객체 이름
          example: clq1234abcd/file5678efgh.jpg
        fileName:
          type: string
          minLength: 1
          description: 원본 파일 이름
          example: image.jpg
        contentType:
          type: string
          minLength: 1
          description: 파일의 MIME 타입
          example: image/jpeg
        fileSize:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: 파일 크기 (bytes)
          example: 102400
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantOption"
          description: (이미지 전용) 생성할 이미지 버전 옵션 배열
      required:
        - fileId
        - objectName
        - fileName
        - contentType
        - fileSize
    FileResponse:
      type: object
      properties:
        id:
          type: string
          description: 파일 ID
          example: file5678efgh
        name:
          type: string
          description: 파일 이름
          example: image.jpg
        mimeType:
          type: string
          description: MIME 타입
          example: image/jpeg
        isImage:
          type: boolean
          description: 이미지 파일 여부
          example: true
        size:
          type: number
          description: 파일 크기 (bytes)
          example: 102400
        url:
          type: string
          nullable: true
          description: 비이미지 파일의 URL
          example: null
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ImageVariantData"
          description: 이미지 버전 목록 (이미지 파일만 해당)
        createdAt:
          type: string
          description: 생성 일시 (ISO 8601)
          example: 2023-01-01T00:00:00.000Z
        updatedAt:
          type: string
          description: 수정 일시 (ISO 8601)
          example: 2023-01-01T00:00:00.000Z
        projectId:
          type: string
          description: 프로젝트 ID
          example: clq1234abcd
      required:
        - id
        - name
        - mimeType
        - isImage
        - size
        - url
        - variants
        - createdAt
        - updatedAt
        - projectId
    ConfirmResponse:
      type: object
      properties:
        message:
          type: string
          description: 성공 메시지
          example: 파일 업로드 완료
        file:
          $ref: "#/components/schemas/FileResponse"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ImageVariantData"
          description: 생성된 이미지 버전 목록
      required:
        - message
        - file
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: 프로젝트 이름 (3-50자)
          example: 내 웹사이트 에셋
        description:
          type: string
          maxLength: 200
          description: 프로젝트 설명 (최대 200자)
          example: 웹사이트에서 사용할 이미지 모음
        storageProvider:
          type: string
          enum:
            - OCI
            - R2
          default: OCI
          description: "스토리지 프로바이더 (기본값: OCI)"
          example: OCI
      required:
        - name
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          description: 프로젝트 ID
          example: clq1234abcd
        name:
          type: string
          description: 프로젝트 이름
          example: 내 웹사이트 에셋
        description:
          type: string
          nullable: true
          description: 프로젝트 설명
          example: 웹사이트에서 사용할 이미지 모음
        storageProvider:
          type: string
          enum:
            - OCI
            - R2
          description: 스토리지 프로바이더
          example: OCI
        createdAt:
          type: string
          description: 생성 일시 (ISO 8601)
          example: 2023-01-01T00:00:00.000Z
        updatedAt:
          type: string
          description: 수정 일시 (ISO 8601)
          example: 2023-01-01T00:00:00.000Z
      required:
        - id
        - name
        - description
        - storageProvider
        - createdAt
        - updatedAt
    ProjectDetailsResponse:
      allOf:
        - $ref: "#/components/schemas/ProjectResponse"
        - type: object
          properties:
            files:
              type: array
              items:
                $ref: "#/components/schemas/FileResponse"
              description: 프로젝트에 속한 파일 목록
          required:
            - files
    ProjectListResponse:
      type: array
      items:
        $ref: "#/components/schemas/ProjectResponse"
  parameters: {}
paths:
  /api/v1/projects:
    get:
      tags:
        - Projects
      summary: 프로젝트 목록 조회
      description: 모든 프로젝트 목록을 조회합니다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 프로젝트 목록이 성공적으로 반환됩니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Projects
      summary: 프로젝트 생성
      description: 새로운 프로젝트를 생성합니다.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: 프로젝트가 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/projects/{projectId}:
    get:
      tags:
        - Projects
      summary: 프로젝트 상세 조회
      description: 프로젝트 ID로 특정 프로젝트와 파일 목록을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: 프로젝트 ID
            example: clq1234abcd
          required: true
          name: projectId
          in: path
      responses:
        "200":
          description: 프로젝트가 성공적으로 반환됩니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetailsResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Projects
      summary: 프로젝트 삭제
      description: 프로젝트 ID로 특정 프로젝트를 삭제합니다. 프로젝트에 속한 모든 파일도 함께 삭제됩니다.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: 프로젝트 ID
            example: clq1234abcd
          required: true
          name: projectId
          in: path
      responses:
        "200":
          description: 프로젝트가 성공적으로 삭제되었습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/projects/{projectId}/files/presign:
    post:
      tags:
        - Files
      summary: Presigned URL 생성
      description: 파일 업로드를 위한 Presigned URL을 생성합니다. 클라이언트는 이 URL을 사용하여 OCI에 직접 파일을 업로드할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: 프로젝트 ID
            example: clq1234abcd
          required: true
          name: projectId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignRequest"
      responses:
        "200":
          description: Presigned URL이 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: 파일 크기 초과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/projects/{projectId}/files/confirm:
    post:
      tags:
        - Files
      summary: 업로드 완료 확인
      description: OCI에 직접 업로드 완료 후 호출하여 DB에 파일 레코드를 생성합니다. 이미지 파일의 경우 variants 옵션으로 변환 처리를 요청할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: 프로젝트 ID
            example: clq1234abcd
          required: true
          name: projectId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "201":
          description: 파일 레코드가 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 파일을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/projects/{projectId}/files/{fileId}:
    delete:
      tags:
        - Files
      summary: 파일 삭제
      description: 특정 ID를 가진 파일 및 관련된 모든 버전을 삭제합니다. (OCI 스토리지 포함)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: 프로젝트 ID
            example: clq1234abcd
          required: true
          name: projectId
          in: path
        - schema:
            type: string
            description: 파일 ID
            example: file5678efgh
          required: true
          name: fileId
          in: path
      responses:
        "200":
          description: 파일 삭제 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 파일을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"